+++ .eslintrc.js
@@ -15,7 +15,7 @@ module.exports = {
   },
   env: {
     browser: true,
-    es6: true,
+    es2020: true,
     node: true,
     jest: true,
   },